Fixed n64 BZN logic.  Fixed minor errors in N64 PrjID map file (note the ODF names are actually too long for BZ1 PC) Added new MalformedArray type, which is a bit of an odd duck but it's used in a few places to show how to use it.  This is needed when invidividual entries of an array can be malformed and we need the ability to fix those per-index on value updates. Other minor fixes like an alternage GetCount function for tokens where the type is passed in for use when using n64 mode files.

+++ BZNParser/Battlezone/EntityDescriptor.cs
@@ -576,7 +576,7 @@ namespace BZNParser.Battlezone
 
             if (writer.Format == BZNFormat.BattlezoneN64)
             {
-                writer.WriteUInt32(null, this, x => x.label, (label) =>
+                writer.WriteUInt16(null, this, x => x.label, (label) =>
                 {
                     if (label != null && label.Value.StartsWith("bzn64label_"))
                     {
